Build and test `aarch64-w64-mingw32` toolchain using MSYS2 MinGW cross-compilation toolchain

+++ .github/scripts/binutils/patch-cygwin.sh
@@ -8,8 +8,10 @@ PATCH_DIR=$SOURCE_PATH/cygwin-packages/binutils
 echo "::group::Patch Cygwin binutils"
     cd $SOURCE_PATH/$BINUTILS_VERSION
 
-    git reset --hard
-    git clean -fdx
+    if [[ "$RESET_SOURCES" = 1 ]]; then
+        git reset --hard HEAD
+        git clean -fdx
+    fi
 
     patch -p2 -i $PATCH_DIR/binutils-2.42-cygwin-config-rpath.patch
     patch -p2 -i $PATCH_DIR/binutils-2.42-cygwin-pep-dll-double-definition.patch
